Enable warnings as error and W4 on MSVC (#83)
This commit increase warning level from `W3` to `W4` on MSVC and sets
warnings as errors. Exclusion is made for a specific warning (C4189).

+++ cli/gcdmodel.cpp
@@ -162,7 +162,8 @@ ErrorCode CGcdData::TranslateToGCD()
     {
         CModelParameter& param = _modelData.Parameters[ index ];
 
-        Parameter* gcdParam = new Parameter( UNDEFINED_ORDER, index, static_cast<int>( param.Values.size() ),
+        Parameter* gcdParam = new Parameter( UNDEFINED_ORDER, static_cast<int>(index),
+                                             static_cast<int>( param.Values.size() ),
                                              param.Name, param.IsResultParameter );
 
         // find out and assign weights to values
